refactor: change folder file & update styles
